Refactor custom integration validation checks
